Trait pallet_offchain_worker::pallet::Config

pub trait Config: Config + Config + CreateSignedTransaction<Call<Self>> {
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Codec + Default + Copy + MaybeSerializeDeserialize;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Call: From<Call<Self>>;
    type AuthorityId: AppCrypto<Self::Public, Self::Signature>;
    type QueryTaskRedundancy: Get<u32>;
    type QuerySessionLength: Get<u32>;
    type Currency: Currency<Self::AccountId>;
    type Reward: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::PositiveImbalance>;
    type OcwQueryReward: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;
    type MaximumWeightForDataAggregation: Get<Weight>;
    type MaximumCommitsPerSession: Get<u32>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Balance: Parameter + Member + AtLeast32BitUnsigned + Codec + Default + Copy + MaybeSerializeDeserialize

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Call: From<Call<Self>>

type AuthorityId: AppCrypto<Self::Public, Self::Signature>

type QueryTaskRedundancy: Get<u32>

type QuerySessionLength: Get<u32>

type Currency: Currency<Self::AccountId>

Currency type for this pallet.

type Reward: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::PositiveImbalance>

Handler for the unbalanced increment when rewarding (minting rewards)

type OcwQueryReward: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

type WeightInfo: WeightInfo

type MaximumWeightForDataAggregation: Get<Weight>

The maximum weight indicate the maximum weight used for data aggregation at the end of a session

type MaximumCommitsPerSession: Get<u32>

The maximum commits per session defined to avoid the weight used for data aggregation exceed MaximumWeightForDataAggregation The weights used is depends on the data size in CommitAccountBalance, as a reference, the weights for 100 items can be found in weights.rs dummy() -> Weight. need re-run the benchmarking according to this parameter and guarantee the data is reasonable