Module pallet_offchain_worker::urls

Structs

EtherScanBalance
Balances data embedded in etherscan response

EtherScanResponse
Response data from etherscan

HttpGet
Http Get URL structure

HttpPost
Http Post URL structure

InfuraBalance
Balances data from Infura service

InfuraResponse
Response from Infura

TokenInfo
Store all API tokens for offchain worker to send request to website

Enums

BlockChainType
Asset type

DataSource
Eth source enum

HttpRequest
Request enum to wrap up both get and post method

Constants

DATA_SOURCE_LIST

TOTAL_DATA_SOURCE_NUMBER

Functions

data_source_to_block_chain_type
Data source to blockchain type

data_source_to_index
Data source to blockchain type

de_string_to_bytes
Deserialize string to Vec

fetch_json_http_get

fetch_json_http_post

get_token

parse_blockchain_info_balances

parse_etherscan_balances

parse_infura_balances

parse_store_tokens

send_get_token